@@ -5,7 +5,7 @@
 	<parent>
 		<groupId>org.springframework.boot</groupId>
 		<artifactId>spring-boot-starter-parent</artifactId>
-		<version>2.7.4</version>
+		<version>3.0.6</version>
 		<relativePath/> <!-- lookup parent from repository -->
 	</parent>
 	<groupId>com.example</groupId>
@@ -14,9 +14,9 @@
 	<name>sales-back</name>
 	<description>Sales backend</description>
 	<properties>
-		<java.version>11</java.version>
-		<spring-cloud.version>2021.0.5</spring-cloud.version>
-		<resilience4j.version>1.7.0</resilience4j.version>
+		<java.version>19</java.version>
+		<spring-cloud.version>2022.0.2</spring-cloud.version>
+		<resilience4j.version>2.0.2</resilience4j.version>
 	</properties>
 	<dependencies>
 		<dependency>
@@ -77,7 +77,7 @@
         </dependency>
 		<dependency>
 			<groupId>io.github.resilience4j</groupId>
-			<artifactId>resilience4j-spring-boot2</artifactId>
+			<artifactId>resilience4j-spring-boot3</artifactId>
 			<version>${resilience4j.version}</version>
 		</dependency>
 		<dependency>
@@ -107,17 +107,19 @@
 		<dependency>
 			<groupId>org.springframework.cloud</groupId>
 			<artifactId>spring-cloud-starter-openfeign</artifactId>
-			<version>2.2.6.RELEASE</version>
+			<version>3.0.6</version>
 		</dependency>
+		<!-- ZIPKING E SLEUTH
 		<dependency>
             <groupId>org.springframework.cloud</groupId>
             <artifactId>spring-cloud-starter-zipkin</artifactId>
-            <version>2.2.8.RELEASE</version>
+            <version>3.0.6</version>
         </dependency>
-		<dependency>
+		
+		 <dependency> 
 			<groupId>org.springframework.cloud</groupId>
 			<artifactId>spring-cloud-starter-sleuth</artifactId>
-		</dependency>
+		</dependency> -->
 	</dependencies>
 	<dependencyManagement>
 		<dependencies>
