@@ -8,7 +8,8 @@ import org.springframework.cloud.openfeign.EnableFeignClients;
 import org.springframework.context.annotation.Bean;
 import org.springframework.web.client.RestTemplate;
 
-import brave.sampler.Sampler;
+//COMMENT: ZIPKIN
+//import brave.sampler.Sampler;
 
 @SpringBootApplication
 @EnableDiscoveryClient
@@ -18,12 +19,12 @@ public class SalesBackApplication {
 	public static void main(String[] args) {
 		SpringApplication.run(SalesBackApplication.class, args);
 	}
-
+/* 
 	@Bean
 	public Sampler defaultSampler(){
 		return Sampler.ALWAYS_SAMPLE;
 	}
-
+*/
 	@LoadBalanced
 	@Bean
 	public RestTemplate getRestTemplate(){
